@@ -55,11 +55,7 @@ namespace PizzaDelivery.ViewModels
                 {
                     return Path.Combine($"\\img\\products\\", NomeArquivoImagem);
                 }
-                else
-                {
-                    return Path.Combine($"\\img\\assets\\sem_imagem.webp");
-
-                }
+                return Path.Combine($"\\img\\assets\\sem_imagem.webp");
             }
         }
 
