@@ -1,6 +1,5 @@
 <!DOCTYPE html>
 <html lang="pt-br">
-
 <head>
     <meta charset="UTF-8">
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
@@ -10,12 +9,15 @@
     <link href="~/css/site.css" rel="stylesheet" />
     <link href="~/lib/bootstrap/css/bootstrap.min.css" rel="stylesheet" />
     <link href="~/lib/bootstrap-icons/font/bootstrap-icons.min.css" rel="stylesheet" />
+
+    <link href="https://fonts.googleapis.com/css?family=Roboto:300,400,500,700&display=swap" rel="stylesheet" />
+
     @await RenderSectionAsync("styles", required: false)
 </head>
 <body>
     <nav class="navbar navbar-expand-lg fixed-top shadow-sm border-bottom bg-gradient">
         <div class="container-fluid">
-            <a class="navbar-brand me-4" href="/"><img src="\img\assets\logo02hor.png" alt="RappiPizzas" width="220px" /></a>
+            <a class="navbar-brand me-4" href="/"><img src="\img\assets\logo02hor.png" alt="RappiPizzas" width="220" /></a>
             <button class="navbar-toggler" type="button" data-bs-toggle="collapse" data-bs-target="#navbarSupportedContent" aria-controls="navbarSupportedContent" aria-expanded="true" aria-label="Toggle navigation">
                 <span class="navbar-toggler-icon"></span>
             </button>
@@ -82,7 +84,7 @@
     </div>
     <!------------------>
 
-    <footer class="text-center glass pt-2">
+    <footer class="text-center glass pt-2 border-light-subtle border-top">
         <div class="d-flex flex-column alig-items-center">
             <div>
                 <img src="\img\assets\logo01vert.png" width="60" alt="RappiPizzas" />
@@ -115,6 +117,9 @@
     <script type="text/javascript" src="~/lib/jquery/jquery.min.js"></script>
     <script type="text/javascript" src="~/lib/bootstrap/js/bootstrap.bundle.min.js"></script>
     <script type="text/javascript" src="~/lib/bootbox.js/bootbox.min.js"></script>
+    <base href="~/" />
+    <script src="_framework/blazor.server.js"></script>
+
     @await RenderSectionAsync("Scripts", required: false)
 
     <script type="text/javascript" src="~/js/site.js" asp-append-version="true" defer></script>
