@@ -1,5 +1,5 @@
 @addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
-@addTagHelper PizzaDelivery.TagHelpers.*, PizzaDelivery
+@* @addTagHelper PizzaDelivery.TagHelpers.*, PizzaDelivery *@
 @using Microsoft.AspNetCore.Identity
 @using PizzaDelivery.Models
 @using PizzaDelivery.ViewModels
